Cleaning up compiler warnings
git-svn-id: https://subversion.nist.gov/JIRA_repo1/STVM_Research/ACVP/gen-val/trunk@388771 58cf25f4-cab2-a041-9f61-d3785993b9bc

+++ common/src/NIST.CVP.Crypto.TDES/KeySchedule.cs
@@ -471,11 +471,11 @@ namespace NIST.CVP.Crypto.TDES
         {
             long output =  input[startIdx] & 0xff;
             output <<= 8;
-            output |= input[startIdx + 1] & 0xff;
+            output |= (long)input[startIdx + 1] & 0xff;
             output <<= 8;
-            output |= input[startIdx + 2] & 0xff;
+            output |= (long)input[startIdx + 2] & 0xff;
             output <<= 8;
-            output |= input[startIdx + 3] & 0xff;
+            output |= (long)input[startIdx + 3] & 0xff;
 
             return output;
         }
